Modify: 컴포넌트에서 호출하던 fetch 함수들을 페이지로 이동
